Module Lib.AssocList

module T : sig ... end

T is a auxiliary module defining an abstract type & its comparison function, following the convention in Jane Street's Base.Compare module

include module type of struct include T end

T is a auxiliary module defining an abstract type & its comparison function, following the convention in Jane Street's Base.Compare module

See https://ocaml.janestreet.com/ocaml-core/v0.12/doc/base/Base/Comparable/index.html

type t = (Base.int * Base.string) Base.list

Definition of the assoc_list opaque type

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val lex_compare_pair : (Base.int * Base.string) -> (Base.int * Base.string) -> Base.int

Polymorphic lexicographic comparison for pair types 'a * 'b, using the default compare functions for types 'a and 'b. For two pairs (a1, b1), (a2, b2):

If compare a1 a2 < 0, lex_compare_pair (a1, _) (a2, _) < 0

If compare a1 a2 > 0, lex_compare_pair (a1, _) (a2, _) > 0

If compare a1 a2 = 0, lex_compare_pair (a1, b1) (a2, b2) = compare b1 b2

Adapted from https://stackoverflow.com/a/20348513

val (>=) : T.t -> T.t -> bool

val (<=) : T.t -> T.t -> bool

val (=) : T.t -> T.t -> bool

val (>) : T.t -> T.t -> bool

val (<) : T.t -> T.t -> bool

val (<>) : T.t -> T.t -> bool

val equal : T.t -> T.t -> bool

val compare : T.t -> T.t -> int

val min : T.t -> T.t -> T.t

val max : T.t -> T.t -> T.t

val ascending : T.t -> T.t -> int

val descending : T.t -> T.t -> int

val between : T.t -> low:T.t -> high:T.t -> bool

val clamp_exn : T.t -> min:T.t -> max:T.t -> T.t

val clamp : T.t -> min:T.t -> max:T.t -> T.t Base__.Or_error.t

type comparator_witness = Base__Comparable.Make(T).comparator_witness

val comparator : ( T.t, comparator_witness ) Base__.Comparator.comparator

val quickcheck_generator : t Base_quickcheck.Generator.t

QuickCheck generator for the assoc_list type

Note: For any user-defined opaque type M.t (where M is a module), we require that the QuickCheck generator be called quickcheck_generator, following the naming conventions of Jane Street's ppx_quickcheck library.