Module AssocList.T

T is a auxiliary module defining an abstract type & its comparison function, following the convention in Jane Street's Base.Compare module

See https://ocaml.janestreet.com/ocaml-core/v0.12/doc/base/Base/Comparable/index.html

type t = (Base.int * Base.string) Base.list

Definition of the assoc_list opaque type

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val lex_compare_pair : (Base.int * Base.string) -> (Base.int * Base.string) -> Base.int

Polymorphic lexicographic comparison for pair types 'a * 'b, using the default compare functions for types 'a and 'b. For two pairs (a1, b1), (a2, b2):

If compare a1 a2 < 0, lex_compare_pair (a1, _) (a2, _) < 0

If compare a1 a2 > 0, lex_compare_pair (a1, _) (a2, _) > 0

If compare a1 a2 = 0, lex_compare_pair (a1, b1) (a2, b2) = compare b1 b2

Adapted from https://stackoverflow.com/a/20348513

val compare : t -> t -> Base.int

Comparison function for the assoc_list opaque type