Module Lib.AssocListMap

module G = Base_quickcheck.Generator

Implementation of maps using association lists.

Adapted from Cornell CS 3110 textbook, chapter 8.

module AssocListMap : MapInterface.MapInterface

Utility functions

val cmp_set_like_lists : 'a list -> 'a list -> bool

cmp_set_like_lists lst1 lst2 compares two association lists to see whether they are equivalent set-like lists. This means:

The two lists can't contain any duplicates

The two lists must contain the same elements though not necessarily in the same order

val pp_string : string -> string

pp_string s pretty-prints the string s

val pp_list : ( 'a -> string ) -> 'a list -> string

pp_list pp_elt lst pretty-prints the list lst, using pp_elt to pretty-print each element of lst

val pp_pair : ( 'a -> string ) -> ( 'b -> string ) -> ('a * 'b) -> string

pp_pair pp1 pp2 (a, b) pretty-prints (a, b) using pp1 for a and pp2 for b