Module Lib.CmdLineParser

Parses user input from the command-line, in particular the filepaths of the .ml files containing the signature & the two modules for testing. (Code adapted from Real World OCaml, Chapter 16)

Utility functions for command-line parsing

val prompt_for_string : string -> ( string -> 'a ) -> 'a

Prompts the user to enter a value, where name is the name of the value, and of_string if a function converting the string input to type 'a

val anon_prompt : string -> ( string -> 'a ) -> 'a Core.Command.Param.t

Creates a parser that automatically prompts the user with the message name if a value isn’t provided, where of_string is used to convert the user input from string to type 'a

val is_filename : string -> string

Checks if a string is a valid readable filename

val regular_file : string Core.Command.Arg_type.t

Defines a cmd-line argument type regular_file which ensures that the input file isn't some Unix file type that can’t be read

Useful filepaths & flags

val pbtFilePath : string

Name of the generated file containing the PBT code

val execFilePath : string

Name of the generated executable file which compares two modules for observational equivalence

val nonNegIntsDoc : string

Documentation string for a flag specifying whether QuickCheck int generators should only generate non-negative ints

val opaqueTypeDoc : string

Docstring for a flag specifying the name of an opaque type contained inside the module signature

val externalLibDoc : string

Docstring for a flag specifying the name of the external library (eg. the Diffie-Hellman x25519 example, see readme)

Writing the generated PBT code to an output file

val writeToPBTFile : ParserTypes.moduleSig -> pbtFilePath:string -> functorName:string -> sigName:string -> externalLib:string option -> nonNegOnly:bool -> string -> string -> unit

Writes the generated PBT code to the file at pbtFilePath m is the parsed module signature AST functorName is the name of the functor that produces the test harness sigName, modName1, modName2 are the names of the signature/two module implementations

Parsing user input from the command-line

val cmdLineParser : Core.Command.t

Parses the names of the signature & implementation files from the cmd-line