Module Lib.CodeGenerator

This file contains the logic for generating PBT code from the AST of a parsed module signature. This file also calls some helper functions defined in the module Lib.Utils.

Functions for generating PBT code

val imports : externalLib:Base.string Base.option -> sigName:Base.string -> modName1:Base.string -> modName2:Base.string -> PPrint.document

imports filepath prints out a PPrint document that imports the requisite modules for the PBT code. The sigName, modName1, modName2 arguments are the names of the module signatures & the two module implementations, which must be the same as their corresponding .ml files.

val sexpAnnotation : PPrint.document

Document for printing the PPX annotation for S-Expr serialization (indented), followed by a newline

val ppxQuickCheckMacro : Base.string -> PPrint.document

Document for invoking the ppx_quickcheck macro for a particular type ty (represented as a string)

val isArrowType : ParserTypes.valDecl -> Base.bool

isArrowType v returns true if the value declaration v has an arrow type

val tyIsArrow : ParserTypes.ty -> Base.bool

tyIsArrow ty returns true if ty correspons to an arrow type

val isOpaqueTy : Base.string -> Base.bool

isOpaqueTy ty returns true if ty is a string corresponding to an "opaque" type, i.e. a type originating from another module

Eg. if ty = "AssocListT", then ty is actually "AssocList.t" where AssocList is some externally-defined module

Note: the only critiera for evaluating if a type is an opaque type is just checking if the type name ends with the substring ".t" (this heuristic may be replaced with a more robust mechanism in the future)

val getIdentityElements : ParserTypes.moduleSig -> Base.string Base.list

Extracts all the "identity elements" defined within a module signature, e.g. mempty for monoids, or zero & one for semirings

For our purposes, we assume any declaration of the form val x : t in the module signature, where t is the module's abstract type, is an identity element

val extractArgTypes : ParserTypes.valDecl -> PPrint.document

Extracts the argument types of functions defined in the module signature, and generates constructors for the expr ADT that take these types as type parameters

val exprADTDecl : ParserTypes.moduleSig -> PPrint.document

Generates the definition of the expr ADT

val printConstructor : Base.string -> Base.string Base.list -> PPrint.document

Helper function for printing out OCaml constructors (Wrapper for the OCaml.variant function in the PPrint library)

val reifyOpaqueTyConstr : Base.string -> Base.string

val splitCamelCaseTyString : Base.string -> Base.string Base.list

val ty_of_string : Base.string -> ParserTypes.ty

ty_of_string s converts a string s denoting a type to the equivalent ty ADT constructor of its base type by running the parser typeP (defined in Lib.ModuleParser) over s.

Note: The string s must be in camel-case

val namedAbsTyVarNameHelper : Base.string -> Base.string

Takes tyName, the name of a named abstract type (eg. public_key), and produces a varname as follows:

Split on underscores

fetch the 1st letter of each substring

Concat all the 1st letters of the substrings

eg. if the type name is public_key, then the varname is pk

val varNameHelper : ParserTypes.ty -> Base.string

varNameHelper ty returns an appropriate variable name corresponding to ty, eg. varNameHelper Int = n

val genVarNameSingleton : ?prime:bool -> ParserTypes.ty -> Base.string

Special case of genVarNames when we only have one argument type. If prime = true, add a single quote to the end of the variable name

val genVarNamesN : n:Base.int -> ParserTypes.ty -> Base.string Base.list

Generates n unique variable names corresponding to a type ty. For example, genVarNamesN 3 Int produces ["n1"; "n2"; "n3"].

val genVarNames : ?prime:bool -> ParserTypes.ty Base.list -> Base.string Base.list

Takes a list of argument types, and generates corresponding variable names which are unique for each element of the list eg. genVarNames [Int; Int; Char] = [n1; n2; c]

val getExprConstructorName : ParserTypes.valDecl -> Base.string

Fetches the expr constructor corresponding to a val declaration in a module

val getExprConstructor : idElts:Base.string Base.list -> ParserTypes.valDecl -> Base.string Base.list * PPrint.document

Fetches the constructor corresponding to a val declaration in the expr ADT, returning a pair of the form (args, constructor applied to args), eg. (["x", "e"], !^ "Mem(x,e)")

The auxiliary argument idElts is a list of names of the identity elements defined in the module signature (eg. mempty for monoids)

See getIdentityElements for further details on identity elements

val extractReturnType : ?camelCase:bool -> ParserTypes.valDecl -> Base.string

extractReturnType camelCase v extracts the string representation of the return type of a function, if v is a valDecl corresponding to an arrow type.

For non-arrow types, this function just extracts the type itself.

The optional argument camelCase determines whether the string returned should be in camelCase or not.

val uniqRetTypesInSig : ?camelCase:bool -> ParserTypes.moduleSig -> Base.string Base.list

Fetches the unique return types across the functions / values in a module signature.

The optional argument camelCase determines whether the string returned should be in camelCase or not.

val tyADTDecl : ParserTypes.moduleSig -> PPrint.document

Generates the definition of the ty ADT

val getFuncName : ParserTypes.valDecl -> PPrint.document

Given a val declaration inside a module (eg. val f : 'a -> 'a), returns the corresponding function name (eg. M.f)

val valADTConstructorString : Base.string -> Base.string

Like valADTConstructor (see below), but returns a string instead of a PPrint document

val valADTConstructor : Base.string -> PPrint.document

valADTConstructor ty generates the constructor name for the value ADT corresponding to the type ty

val tyADTConstructors : ?camelCase:bool -> ParserTypes.moduleSig -> Base.string Base.list

Generates the list of constructor names (strings) for the ty ADT

The optional argument camelCase determines whether the strings returned should be in camelCase or not.

val tyAndValADTConstructors : ?camelCase:bool -> ParserTypes.moduleSig -> (Base.string * Base.string) Base.list

Returns an association list of string pairs, where each element is a constructor for the ty ADT & its associated constructor for the value ADT.

The optional argument camelCase determines whether the strings returned should be in camelCase or not.

val instantiateT : Base.string -> Base.String.t

Takes s, a string reprentation of a type, and instantiates the abstract type t to int M.t where M is some module defined elsewhere.

Example: instantiate "toption" = "int M.t option"

If s doesn't contain the prefix "t", s is left unchanged

val addSpaceToTyStr : Base.string -> Base.String.t

Takes s, a string reprentation of a type, and adds a space to s if s represents a parameterized type.

Example: addSpaceToTyStr "intoption" = "int option"

If s doesn't contain the suffix "list" or "option", s is left unchanged.

val destructAbsTy : ParserTypes.abstractType -> PPrint.document

destructAbsTy takes an abstract type absTy and produces a PPrint document referencing absTy inside some module M. For example, the abstract type t (represented as the inhabitant T0 "t" of the type abstractType) corresponds to "M.t".

Note: any polymorphic type variables are automatically instantiated with int, so 'a t becomes "int M.t".

val lookupAbsTy : Base.string -> ParserTypes.abstractType Base.list -> ParserTypes.abstractType option

Takes a list absTys of abstract types and checks if the string s is the name of an abstract type

val valADTParam : moduleAbsTys:ParserTypes.abstractType Base.list -> ?isOpaque:bool -> Base.string -> PPrint.document

valADTParam moduleAbsTy ty takes ty, a string representation of the corresponding type, and generates the type parameter for the constructor value ADT corresponding to the type ty.

The auxiliary argument moduleAbsTy refers to the abstract type contained within the module signature, e.g. M.t

The optional argument isOpaque is a boolean indicating whether this type is "opaque" (i.e. originating from another module, eg. "AssocList.t" where "AssocList" is an externally-defined module)

If the module's abstract type M.t is polymorphic (i.e. 'a M.t), we instantiate 'a with int, otherwise we leave the abstract type monomorphic (i.e. just M.t)

val valADTTypeDef : moduleAbsTys:ParserTypes.abstractType Base.list -> Base.string -> PPrint.document

valADTTypeDef moduleAbsTy ty generates both the constructor & type parameter for the value ADT corresponding to the type ty

The auxiliary argument moduleAbsTy is the abstract type defined in the module, which is used to determine if the abstract type needs to be instantiated with a concrete type (this logic is handled in valADTParam)

Note: This function is a helper function called by valADTDefn

val valADTDefn : ParserTypes.moduleSig -> PPrint.document

Generates the value ADT definition that is contained within the module returned by the ExprToImpl functor

val interpIsNeeded : ParserTypes.ty -> Base.bool

Given an argument and its type, determines if we need to recursively call interp on the argument for the inner pattern match in interp

type argPos =

| Fst

| Snd

Auxiliary data type for indicating the position of a non-expr argument to a function

val argPos_of_sexp : Sexplib0.Sexp.t -> argPos

val sexp_of_argPos : argPos -> Sexplib0.Sexp.t

val interpOnce : ParserTypes.ty -> ?nonExprArg:(string * argPos) option -> PPrint.document -> ParserTypes.ident -> ParserTypes.ty -> PPrint.document

Pattern matches interp on one argument of type expr. If nonExprArg is Some of some value, it is placed in the appropriate argument position during function application.

val interpTwice : ParserTypes.ty -> ParserTypes.ty -> PPrint.document -> ParserTypes.ident -> ParserTypes.ident -> ParserTypes.ty -> PPrint.document

Pattern matches interp on two arguments, both of type expr

val interpExprPatternMatch : (ParserTypes.valDecl * Base.string Base.list) -> PPrint.document

Produces the inner pattern match (interp e) in the interp function

val interpDefn : ParserTypes.moduleSig -> PPrint.document

interpDefn m generates the definition of the interp function, which evaluates exprs based on the value declarations defined in the module signature m. This function does so by doing the following:

Extracting the names of all the identity elements in m by calling getIdentityElements

Fetching the expr ADT constructors & arguments based on the declarations inside m

Generating the pattern matches inside the interp function

val functorDef : ParserTypes.moduleSig -> sigName:Base.string -> functorName:Base.string -> PPrint.document

Generates the definition of the ExprToImpl functor

val getGenerator : ?nonNegOnly:bool -> ParserTypes.ty -> PPrint.document

getGenerator ty takes in a type ty and produces a PPrint document containing the corresponding QuickCheck generator for that type.

The auxiliary argument nonNegOnly specifies whether QuickCheck's int generators should only generate non-negative ints (in the event ty is Int).

By default, the int generator returned generates both negative & positive ints.

Note that polymorphic types (eg. Alpha) are instantiated as ints

This is a helper function called by argGen.

val argGen : ?nonNegOnly:bool -> ?pairName:Base.string Base.Option.t -> Base.string -> ParserTypes.ty -> PPrint.document

Produces the code for a monadic bind of arg to a QuickCheck generator producing a value of type ty, where ty must be a non-arrow type.

This is a helper function called by genExprPatternRHS.

The optional argument nonNegOnly specifies whether QuickCheck's int generators should only generate non-negative ints

If the optional argument pairName (of type string option) is Some p and ty is of the form Pair (_, _), p will be used as the variable name for the overall pair.

val isPairType : ParserTypes.ty -> Base.bool

isPairType ty returns true if ty is a pair type, false otherwise

val genExprPatternRHS : ?nonNegOnly:bool -> (Base.string * ParserTypes.ty * Base.string Base.list * Base.string * PPrint.document) -> Base.string * (ParserTypes.ty * PPrint.document * Base.string)

Takes in the following arguments:

nonNegOnly: An optional argument specifying whether QuickCheck's int generators should only generate non-negative ints

tyConstr: constructor for the ty ADT representing the return type of a function

funcTy: function type

args: arguments passed to the Expr constructor

constr: constructor for the Expr ADT

funcApp: PPrint document containing the application of constr onto args

This function produces the RHS of the pattern matches in gen_expr, returning a tuple of the form (tyConstr, (ty, patternMatchRHS, nameOfPatternMatch))

For example, if:

tyConstr = Bool

ty = Func2(Alpha, Expr, Bool)

constr = Mem

funcApp = Mem(x, e) then genExprPatternRHS (tyConstr, ty, args, constr, funcApp) produces the code

let%bind arg = G.int in G.return @@ Mem(x, e)

where G = Base_quickcheck.Generator.

val getExprConstructorWithArgs : Base.string -> ParserTypes.ty -> Base.string -> Base.string * ParserTypes.ty * Base.string Base.list * Base.string * PPrint.document

Arguments:

tyConstr (constructor for the ty ADT representing the return type of a function)

ty (the type of the Expr construcctor)

constr (the string representation of an Expr constructor)

Returns: a 5-tuple of the form

(tyConstr, ty, constructorArgs, constructor, constructor applied to args) 

e.g.

(Bool, Func2(Alpha, Expr, Bool), ["x", "e"], "Mem", !^ "Mem(x,e)") 

val genExprPatterns : ParserTypes.moduleSig -> (Base.string * (ParserTypes.ty * PPrint.document * Base.string)) Base.List.t

Returns an association list of constructors for the ty ADT where each element is the form (<constructor for the ty ADT>, ty)

val setNonNegIntFlag : ParserTypes.moduleSig -> ParserTypes.moduleSig

setNonNegIntFlag m sets the intFlag field of the module signature m to NonNegativeOnly, indicating that QuickCheck's int generators should only generate non-negative ints.

val genExprDef : nonNegOnly:Base.bool -> ParserTypes.moduleSig -> PPrint.document

Produces the definition of the gen_expr Quickcheck generator for the expr datatype.

The auxiliary argument nonNegOnly is a boolean specifying if QuickCheck's integer generators should only generate non-negative integers. (This is an optional command-line flag that is passed in when Mica is invoked.)

An example of the automatically-produced code for gen_expr can be found in GeneratedSetPBTCode.ml (specialized to the finite set example discussed in the README / documentation homepage).

When the internal size parameter of gen_expr reaches 0, the function yields a trivial generator that just returns the identity element (or one of them, if there are multiple) of the signature.

See getIdentityElements for a further discussion on identity elements.

val functorApp : functorName:ParserTypes.moduleName -> Base.string -> PPrint.document

Produces a PPrint document representing an OCaml functor application, where functorName is the name of the functor, and arg is the name of the argument to the functor

val implModuleBindings : functorName:ParserTypes.moduleName -> Base.string -> Base.string -> PPrint.document

Produces the module bindings for the two module implemntations I1, I2, where functorName is the name of the functor that procduces the PBT test harness, and modName1 & modName2 are the names of the two module implementaitons

Functions for generating the executable for comparing two modules

val displayErrorDef : PPrint.document

Generates the definition of a displayError helper function for displaying error messages when QuickCheck tests fail

val executableImports : pbtFilePath:Base.string -> execFilePath:Base.string -> PPrint.document

Generate requisite imports for the executable file, where filepath is the filepath to the executable file

val getTestName : Base.string -> PPrint.document

Given tyConstr, a constructor in the ty datatype to an OCaml datatype, getTestName tyConstr produces an interstitial variable name corresponding to this type with the prefix "test_", in the form of a PPrint document

For example, getTestName Int produces "test_int" (as a PPrint document)

val obsEquiv : (Base.string * Base.string) -> PPrint.document

Produces Quickcheck code in the executable that tests two modules for observational equivalence based on exprs that return some tyConstr of type ty, and pattern matching on the equivalent valConstrs that they return

val isExcludedType : ParserTypes.ty -> Base.bool

isExcludedType ty returns true if the type ty is a return type to be excluded from consideration when testing for observational equivalence.

For example, expressions returning abstract types T and AlphaT are excluded when we check if two modules for observational equivalence, because these abstract types may be instantiated differently in the two modules.

For instance, one module could instantiate 'a t to be 'a list, while the other instantiates 'a t to be 'a tree.

Note: we cannot test opaque types and named abstract types directly for observational equivalence (due to encapsulation, we can't inspect these types directly)

Note that arrow types are not considered by this function.

val compareImpls : ParserTypes.moduleSig -> PPrint.document

Generate the executable code for testing observational equivalence of two modules