Module Lib.DFA

Representation of DFAs in OCaml

Adapted from Harry Goldstein's video on DFAs & regexes

String utility functions

val string_of_char : Base.char -> Base.string

Converts a char to a string

val explode : Base.string -> Base.char Base.list

explode str converts the string str to a list of chars

val implode : Base.char Base.list -> Base.string

implode chars converts chars, a list of chars, to a string

DFA representation

type dfa = {

acceptsEmpty : Base.bool;

next : Base.char -> dfa;

}

val next : dfa -> Base.char -> dfa

val acceptsEmpty : dfa -> Base.bool

module Fields_of_dfa : sig ... end

val dfa_of_sexp : Sexplib0.Sexp.t -> dfa

val sexp_of_dfa : dfa -> Sexplib0.Sexp.t

module DFA : RegexMatcher.RegexMatcher

DEPRECATED: convert re converts the regex re to a DFA