Module type DirectAddressMap.DirectAddressMap

Signature for mutable maps with integer keys Adapted from the Cornell CS 3110 textbook, chapter 8.

type 'v t

'v t is the type of maps that binds keys of type int to values of type 'v

val insert : int -> 'v -> 'v t -> unit

insert k v m mutates map m to bind k to v. If k was already bound in m, that binding is replaced by the binding to v in the new map.

val find : int -> 'v t -> 'v option

find k m is Some v if k is bound to v in m, and None if not.

val remove : int -> 'v t -> unit

remove k m mutates m to remove any binding of k. If k was not bound in m, then the map is unchanged.

val empty : 'v t

empty is the empty map

val of_list : (int * 'v) list -> 'v t

of_list lst is a map containing the same bindings as the association list lst. Requires: lst does not contain any duplicate keys.

val bindings : 'v t -> (int * 'v) list

bindings m is an association list containing the same bindings as m. There are no duplicate keys in the list.