Module Lib.GeneratedRegexPBTCode

type expr =

| Void

| Empty

| Lit of Base.char

| Alt of expr * expr

| Cat of expr * expr

| Star of expr

| MatchString of expr * Base.string

| AcceptsEmpty of expr

val expr_of_sexp : Sexplib0.Sexp.t -> expr

val sexp_of_expr : expr -> Sexplib0.Sexp.t

val compare_expr : expr -> expr -> Base.int

val hash_fold_expr : Ppx_hash_lib.Std.Hash.state -> expr -> Ppx_hash_lib.Std.Hash.state

val hash_expr : expr -> Ppx_hash_lib.Std.Hash.hash_value

val normalize : expr -> expr

"Shrinker" for regexes

type ty =

| Bool

| T

val ty_of_sexp : Sexplib0.Sexp.t -> ty

val sexp_of_ty : ty -> Sexplib0.Sexp.t

module ExprToImpl (M : RegexMatcher.RegexMatcher) : sig ... end

val gen_expr : ty -> expr Base_quickcheck.Generator.t

module I1 : sig ... end

module I2 : sig ... end

val displayError : expr -> I1.value -> I2.value -> Base.string

type observableExpr = [

| `MatchString

| `AcceptsEmpty

]

The type of exprs that produce observable values

Note that for this interface, only the expr constructors MatchString and AcceptsEmpty produce observable values (bools)

val __observableExpr_of_sexp__ : Sexplib0.Sexp.t -> observableExpr

val observableExpr_of_sexp : Sexplib0.Sexp.t -> observableExpr

val sexp_of_observableExpr : observableExpr -> Sexplib0.Sexp.t

val compare_observableExpr : observableExpr -> observableExpr -> Base.int

val hash_fold_observableExpr : Ppx_hash_lib.Std.Hash.state -> observableExpr -> Ppx_hash_lib.Std.Hash.state

val hash_observableExpr : observableExpr -> Ppx_hash_lib.Std.Hash.hash_value

val last : expr -> observableExpr

Picks out the last constructor invoked in an expr value, returning an observableExpr

module ObservableExpr : sig ... end

Auxiliary module that derives comparison and hash functions for the type observableExpr

module ExprBool : sig ... end

Auxiliary module that derives comparison and hash functions for the type observableExpr * bool * bool