Module type HashTable.HashTableInf

Module signature / interface for hash tables

type ('k, 'v) t

('k, 'v) t is the type of mutable table-based maps that bind keys of type 'k to values of type 'v.

val insert : 'k -> 'v -> ( 'k, 'v ) t -> unit

insert k v m mutates map m to bind k to v. If k was already bound in m, that binding is replaced by the binding to v.

val find : 'k -> ( 'k, 'v ) t -> 'v option

find k m is Some v if m binds k to v, and None if m does not bind k.

val remove : 'k -> ( 'k, 'v ) t -> unit

remove k m mutates m to remove any binding of k. If k was not bound in m, the map is unchanged.

val create : ( 'k -> int ) -> int -> ( 'k, 'v ) t

create hash c creates a new table map with capacity c that will use hash as the function to convert keys to integers. Requires: The output of hash is always non-negative, and hash runs in constant time.

val bindings : ( 'k, 'v ) t -> ('k * 'v) list

bindings m is an association list containing the same bindings as m.

val of_list : ( 'k -> int ) -> ('k * 'v) list -> ( 'k, 'v ) t

of_list hash lst creates a map with the same bindings as lst, using hash as the hash function. Requires: lst does not contain any duplicate keys.