Module HashTableTest.HConf

type state = (string * int) list

For simplicity, we test hashtables with string keys & int values

type sut = ( string, int ) Core.Hashtbl.t

For simplicity, we test hashtables with string keys & int values

type cmd =

| Clear

| Add of string * int

| Remove of string

| Find of string

| Replace of string * int

| Mem of string

| Length

val sexp_of_cmd : cmd -> Sexplib0.Sexp.t

val quickcheck_generator_cmd : cmd Ppx_quickcheck_runtime.Base_quickcheck.Generator.t

val quickcheck_observer_cmd : cmd Ppx_quickcheck_runtime.Base_quickcheck.Observer.t

val quickcheck_shrinker_cmd : cmd Ppx_quickcheck_runtime.Base_quickcheck.Shrinker.t

val gen_cmd : state -> cmd Core.Quickcheck.Generator.t

val init_state : 'a list

Initial state for the model is just an empty list

val next_state : cmd -> ( Core.String.t, int ) Core.List.Assoc.t -> ( Core.String.t, int ) Core.List.Assoc.t

val init_sut : unit -> ( Core.String.t, 'a ) Core.Hashtbl.t

Initialize a Hashtable with String keys w/ initial size 42

val cleanup : 'a -> unit

val run_cmd : cmd -> ( Core.String.t, int ) Core.List.Assoc.t -> ( string, int ) Core.Hashtbl.t -> bool

val precond : cmd -> state -> bool