Module type MapInterface.MapInterface

type t

The abstract type of maps. For now, we require keys to be int and values to be strings.

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val empty : t

empty is the empty map

val insert : (int * string) -> t -> t

insert (k, v) m produces a new map that is the same as m, but with an additional binding from k to v. If k was already bound in m, that binding is replaced by the binding to v in the new map.

val find : int -> t -> string option

find k m is Some v if k is bound to v in m, and None if not.

val remove : int -> t -> t

remove k m produces a new map that is the same as m, but with all the bindings for k removed. If k was not bound in m, then the map is unchanged.

val from_list : AssocList.t -> t

from_list lst is a map containing the same bindings as the association list lst. Requirement: lst does not contain any duplicate keys.

val bindings : t -> AssocList.t

bindings m is an association list containing the same bindings as m. There are no duplicate keys in the list.