Module Lib.ModuleParser

A parser for module signatures, based on the Angstrom parser-combinator library.

This file depends on helper functions defined in the module Lib.Parser, and datatype definitions defined in Lib.ParserTypes.

module A = Angstrom

Alias for the Angstrom module

val identP : ?firstCharP:Base.char A.t -> unit -> Base.string A.t

Parsers for identifiers in OCaml ident ::= (A..Z | a..z | _) {A..Z | a..z | 0..9 | _ | \'}

val lowercaseIdentP : Base.string A.t

Parser for identifiers which start with lowercase letters / an underscore

val typeParamP : Base.char A.t

Parses type-parameters, eg. 'a

val sexpAnnotP : Base.unit A.t

Parses the S-expression PPX annotation, consuming subsequent whitespace TODO: this parser could be written more cleanly

val sigP : 'a A.t -> 'a A.t

sigP p takes a parser p, and sandwiches it between parsers that parse the "sig" and "end" tokens

val modNameP : Base.string A.t

Parser for module names, which must start with a capital letter

val baseTypeP : ParserTypes.ty A.t

Parser for base types

val libraryNameP : Base.string A.t

Parses the names of common Jane Street libraries

val paramTypeP : ParserTypes.ty A.t

General parser for parameterized types, eg. options/pairs/lists

val opaqueTypeP : ParserTypes.ty A.t

Parser for "opaque types", i.e. types that are sealed within another module, eg. the type M.t from another module M

val namedAbstractTypeP : ParserTypes.ty A.t

Parser for the NamedAbstract case of the ty datatype, i.e. parser for named monomorphic abstract types, e.g. type private_key in the Diffie-Hellman example

val typeP : ParserTypes.ty A.t

General parser for non-arrow types

val typeTokenP : Base.unit A.t

Parses the token "type"

val abstractTypeP : ParserTypes.abstractType A.t

Parser for the declaration of an abstract type in a module, eg. type 'a t

val arrowTypeP : ParserTypes.ty A.t

Parser for arrow types

val valDeclP : ParserTypes.valDecl A.t

Parser for a value declaration inside a module, eg. val empty : 'a t or val func : 'a -> 'a t -> 'a t

val moduleTypeP : ParserTypes.moduleSig A.t

Parser for a module signature

val string_of_file : Base.string -> Base.string

Reads a file as a single string

val file_of_string : Base.string -> Base.string -> Base.unit

Writes a string to a file