Module Lib.ParserTypes

Datatype definitions for OCaml module signatures. These types are used by the module Lib.ModuleParser to parse OCaml module signatures. This file also calls some helper functions defined in the module Lib.Utils.

type moduleName = Base.string

Name of a module

val moduleName_of_sexp : Sexplib0.Sexp.t -> moduleName

val sexp_of_moduleName : moduleName -> Sexplib0.Sexp.t

type moduleType =

| Intf

| Impl

A module can either be an interface Intf or an implementation Impl

val moduleType_of_sexp : Sexplib0.Sexp.t -> moduleType

val sexp_of_moduleType : moduleType -> Sexplib0.Sexp.t

type ident = Base.string

Identifiers in OCaml

val ident_of_sexp : Sexplib0.Sexp.t -> ident

val sexp_of_ident : ident -> Sexplib0.Sexp.t

type ty =

| Int

| Char

| Bool

| Unit

| String

| Alpha

| AlphaT

| T

| NamedAbstract of Base.string

| Opaque of Base.string

| Option of ty

| Pair of ty * ty

| List of ty

| Func1 of ty * ty

| Func2 of ty * ty * ty

ty is an ADT representing concrete types, where:

Alpha represents 'a

AlphaT represents 'a t

T represents t

Option represents option types

Pair represents a pair, i.e. a product type with two arguments

List represents lists

Func1 represents functions of arity 1 (arg type, return type)

Func2 represents functions of arity 2 (arg1 type, arg2 type, return type)

NamedAbstract represents a named monomorphic abstract type, i.e. an abstact type that is not t, e.g. the type private_key in the Diffie-Hellman example. Note: polymorphic named abstract types are currently not supported.

Opaque represents an opaque type defined in an external module. An example of an opaque type is AssocList.t, where AssocList is an externally defined module (defined in a file AssocList.ml in the lib directory). We require that the type in this external module be called t, and that the QuickCheck generator should be called quickcheck_generator, following the naming conventions of Jane Street's ppx_quickcheck library.

See lib/maps/AssocList.ml for an example of opaque types in use.

All other constructors correspond to base OCaml types

val ty_of_sexp : Sexplib0.Sexp.t -> ty

val sexp_of_ty : ty -> Sexplib0.Sexp.t

val compare_ty : ty -> ty -> Base.int

type abstractType =

| T0 of Base.string

| T1 of ty * Base.string

Abstract type contained within a module

The T0 constructor means the abstract type is not parameterized by any type variables, i.e. the abstract type is just t

The string argument for the T0 constructor is the name of the abstract type (eg. "t" or "private_key" for the Diffie-Hellman example)

The T1 constructor takes a type ty as its argument, representing an abstract type containing a type variable, eg. 'a t

T1 also takes in a string argument, which is the name of the abstract type (eg. "t")

val abstractType_of_sexp : Sexplib0.Sexp.t -> abstractType

val sexp_of_abstractType : abstractType -> Sexplib0.Sexp.t

val tyEqual : ty -> ty -> Base.bool

Structural equality function for the ty datatype

val string_of_ty : ?alpha:Base.string -> ?t:Base.string -> ?camelCase:bool -> ty -> Base.string

Converts a ty to its string representation, taking in optional arguments specifying various parameters of the string represntation.

Note that AlphaT & T are converted to "expr"

The argument alpha specifies a concrete base type that should be printed in the PBT code in lieu of \'a when testing polymorphic functions.

The argument t specifies a string that should be printed whenever ty is equal to AlphaT or T (eg. the string literal "expr" when generating the expr ADT definition)

The argument camelCase specifies whether the string representation of ty should be in camel-case or not. This is relevant for parameterized types whose string representation may contain a space, eg. "int option".

type valDecl = {

valName : Base.string;

valType : ty;

}

Type representing a value declaration:

valName is the name of the variable/function being declared

valType is the type of the declaration

E.g. for val empty : 'a t, valName = "empty" and valType = AlphaT

val valDecl_of_sexp : Sexplib0.Sexp.t -> valDecl

val sexp_of_valDecl : valDecl -> Sexplib0.Sexp.t

val valType : valDecl -> ty

val valName : valDecl -> Base.string

module Fields_of_valDecl : sig ... end

type intFlag =

| AllInts

| NonNegativeOnly

| PositiveOnly

Auxiliary type which specifies whether integer QuickCheck generators should generate:

AllInts (i.e. all positive/negative ints)

NonNegativeOnly (only generate non-negative ints)

PositiveOnly (only generate strictly positive ints)

Note: when a module signature is parsed in the ModuleParser module, by default, intFlag is set to AllInts

val intFlag_of_sexp : Sexplib0.Sexp.t -> intFlag

val sexp_of_intFlag : intFlag -> Sexplib0.Sexp.t

type moduleSig = {

moduleName : moduleName;

moduleType : moduleType;

abstractTypes : abstractType Base.list;

valDecls : valDecl Base.list;

intFlag : intFlag;

}

Record type representing an ML module

val moduleSig_of_sexp : Sexplib0.Sexp.t -> moduleSig

val sexp_of_moduleSig : moduleSig -> Sexplib0.Sexp.t

val intFlag : moduleSig -> intFlag

val valDecls : moduleSig -> valDecl Base.list

val abstractTypes : moduleSig -> abstractType Base.list

val moduleType : moduleSig -> moduleType

val moduleName : moduleSig -> moduleName

module Fields_of_moduleSig : sig ... end