Module Lib.Stats

Utility functions for printing out distributions of reified values over the modules under test.

Note: This module is intended to be opened before the file containing the auto-generated PBT code is opened. Note that the type expr is declared abstractly in the file, whereas expr is instantiated with a concrete definition in the auto-generated files produced by Mica (e.g. lib/generated.ml). Opening this file prior to opening lib/generated.ml will allow these utility functions to operate on the concrete definition of the expr datatype.

See bin/GeneratedSetExecutable.ml for an example of how to invoke the functions in this file

type expr

Abstract type of symbolic expressions

val expr_of_sexp : Sexplib0.Sexp.t -> expr

val sexp_of_expr : expr -> Sexplib0.Sexp.t

val compare_expr : expr -> expr -> int

module Expr : sig ... end

Module containing a comparator witness & serialization functions for the expr datatype. This module is necessary for creating maps/sets over exprs -- see Chapter 14 of Real World OCaml for further details on this programming idiom. (Also need to update the appropriate PPX macro invocations in GeneratedSetPBTCode.ml)

val seq_coverage : expr Core.Sequence.t -> int Core.Map.M(Expr).t

Given a sequence seq of generated exprs, seq_coverage seq produces a Map which maps each unique expr to its frequency in seq

val genSeqPrintFreq : expr Base_quickcheck.Generator.t -> unit

genSeqPrintFreq gen takes a generator gen of exprs, uses gen to produce a random sequence of exprs, and prints the frequency of each expr in the random sequence to stdout

val incrFreq : ( 'a, int, 'comparator ) Core.Map.t -> 'a -> ( 'a, int, 'comparator ) Core.Map.t

incrFreq m key increments the frequency of key in the map m

val incrHashtbl : ( 'a, int ) Core.Hashtbl.t -> 'a -> unit

Increments the value of a key in a Hashtbl mapping 'a values to int

val genExprWithLog : expr Base_quickcheck.Generator.t -> ( expr, int ) Core.Hashtbl.t -> expr Base_quickcheck.Generator.t

genExprWithLog gen h takes a generator gen of exprs, and produces a new generator that generates the same values but also increments a hashtable h of frequencies (mapping expr to int) every time a value is generated

val sumValues : ( 'a, int ) Core.Hashtbl.t -> int

Computes the sum of all the values in the hashmap h

val computePercents : ( 'a, int ) Core.Hashtbl.t -> ( 'a, float ) Core.Hashtbl.t

Takes a hashtable mapping 'a keys to their occurrences, and converts it into a hashtable mapping 'a to their frequency (as a percentage)

val sortByPercent : ( 'a, int ) Core.Hashtbl.t -> (float * 'a) list

Takes a hashtable mapping 'a keys to their integer frequency, computes the frequency of each key as a percentage, then returns a float * 'a list association list sorted in descending order of percentage (highest frequency comes first)

val printPercents : ?precision:int -> (float * 'a) -> (float * 'a) -> printKey:( 'a -> string ) -> unit

Prints the percentages of two pairs, each of type float * 'a list

The argument ~printKey converts type 'a to string for printing

The argument ~precision specifies the no. of characters to print in the string representation of the 'a type. The default precision is 10.

val printPercent : ?precision:int -> (float * 'a) -> printKey:( 'a -> string ) -> unit

Prints a percentage of a pair of type float * 'a, where the percentage is the first element of the pair

See printPercents for explanations of the named arguments precision and printKey

Prints the percentage of each key's occurrence in the hashtable h, using printKey as the serialization for keys of type 'a