Module Lib.TestHarness

module type SpecBase = sig ... end

Base signature that the PBT harness must satisfy

module type Spec1 = sig ... end

Extension of SpecBase for testing one module wrt a signature

module type Spec2 = sig ... end

Extension of SpecBase for comparing two modules that implement the same interface

module type HarnessBase = sig ... end

Base module type for the PBT harness

module type Harness1 = sig ... end

Extension of HarnessBase, specialized for testing one module

module type Harness2 = sig ... end

Extension of HarnessBase, specialized for comparing two modules

module Make1 (Spec : Spec1) : Harness1 with type state := Spec.state and type cmd := Spec.cmd and type sut := Spec.sut

Given a PBT specification, the functor Make1 creates a test harness that tests if a module implements its signature properly

module Make2 (Spec : Spec2) : Harness2 with type state := Spec.state and type cmd := Spec.cmd and type sutA := Spec.sutA and type sutB := Spec.sutB

Given a PBT specification, the functor Make2 creates a test harness that compares if two modules implement the same signature