Module TestHarness.Make1

Given a PBT specification, the functor Make1 creates a test harness that tests if a module implements its signature properly

Parameters

module Spec : Spec1

Signature

include HarnessBase with type state := Spec.state with type cmd := Spec.cmd

val gen_cmds : Spec.state -> int -> Spec.cmd list Base_quickcheck.Generator.t

A fueled command list generator. Accepts a state parameter to enable state-dependent cmd generation.

val cmds_ok : Spec.state -> Spec.cmd list -> bool

A precondition checker (stops early, thanks to short-circuit Boolean evaluation). Accepts the initial state and the command sequence as parameters.

val arb_cmds : Spec.state -> Spec.cmd list Base_quickcheck.Generator.t

A generator of command sequences. Accepts the initial state as parameter.

val consistency_test : trials:int -> unit

A consistency test that generates a number of cmd sequences and checks that all contained cmds satisfy the precondition precond. Accepts a labeled parameters trials, which is the no. of trials

val interp_agree : Spec.state -> Spec.sut -> Spec.cmd list -> bool

Checks agreement between the model and the system under test (stops early, thanks to short-circuit Boolean evaluation).

val agree_prop : Spec.cmd list -> bool

Agreement property: checks if the command sequence cs yields the same observations when interpreted from the model's initial state and the sut's initial state. Cleans up after itself by calling relevant functions in other modules.

val agree_test : int -> unit

An actual agreement test (for convenience). Accepts an optional count parameter * and a test name as a labeled parameter name.