Module type TestHarness.Harness2

Extension of HarnessBase, specialized for comparing two modules

include HarnessBase

type state

type cmd

val gen_cmds : state -> int -> cmd list Base_quickcheck.Generator.t

A fueled command list generator. Accepts a state parameter to enable state-dependent cmd generation.

val cmds_ok : state -> cmd list -> bool

A precondition checker (stops early, thanks to short-circuit Boolean evaluation). Accepts the initial state and the command sequence as parameters.

val arb_cmds : state -> cmd list Base_quickcheck.Generator.t

A generator of command sequences. Accepts the initial state as parameter.

val consistency_test : trials:int -> unit

A consistency test that generates a number of cmd sequences and checks that all contained cmds satisfy the precondition precond. Accepts a labeled parameters trials, which is the no. of trials

type sutA

Types of the two systems under tests (SUTs)

type sutB

val interp_agree : state -> sutA -> sutB -> cmd list -> bool

Checks agreement between the model and the two SUTs