Module type TestHarness.Spec1

Extension of SpecBase for testing one module wrt a signature

include SpecBase

type cmd

The type of commands

type state

The type of the model's state

val gen_cmd : state -> cmd Core.Quickcheck.Generator.t

A command generator. Accepts a state parameter to enable state-dependent cmd generation.

val init_state : state

The model's initial state.

val next_state : cmd -> state -> state

next_state c s expresses how interpreting the command c moves the model's internal state machine from the state s to the next state. Ideally a next_state function is pure.

val precond : cmd -> state -> bool

precond c s expresses preconditions for command c in terms of the model state s. A precond function should be pure. precond is useful, e.g., to prevent the shrinker from breaking invariants when minimizing counterexamples.

type sut

The type of the system under test

val init_sut : unit -> sut

Initialize the system under test.

val cleanup : sut -> unit

Utility function to clean up the sut after each test instance, e.g., for closing sockets, files, or resetting global parameters

val run_cmd : cmd -> state -> sut -> bool

run_cmd c s i should interpret the command c over the system under test (typically side-effecting). s is in this case the model's state prior to command execution. The returned Boolean value should indicate whether the interpretation went well and in case c returns a value: whether the returned value agrees with the model's result.