Module type TestHarness.Spec2

Extension of SpecBase for comparing two modules that implement the same interface

include SpecBase

type cmd

The type of commands

type state

The type of the model's state

val gen_cmd : state -> cmd Core.Quickcheck.Generator.t

A command generator. Accepts a state parameter to enable state-dependent cmd generation.

val init_state : state

The model's initial state.

val next_state : cmd -> state -> state

next_state c s expresses how interpreting the command c moves the model's internal state machine from the state s to the next state. Ideally a next_state function is pure.

val precond : cmd -> state -> bool

precond c s expresses preconditions for command c in terms of the model state s. A precond function should be pure. precond is useful, e.g., to prevent the shrinker from breaking invariants when minimizing counterexamples.

type sutA

Types of the two systems under tests (SUTs)

type sutB

val init_sutA : unit -> sutA

Initialization functions for the two SUTs

val init_sutB : unit -> sutB

val cleanupA : sutA -> unit

Cleanup functions for the two SUTs

val cleanupB : sutB -> unit

val compare_cmd : cmd -> state -> sutA -> sutB -> bool

compare_cmd cmd st sutA sutB interprets the symbolic command cmd over the two SUTs, and checks whether the result of interpreting cmd over sutA & sutB agree with each other. Here, st refers to the model's state prior to executing cmd.